requirements.txt
tensorflow_gpu==1.10.0
numpy==1.13.3
opencv_python==3.2.0.7
matplotlib==2.2.2
easydict==1.6
flask==1.0.2
pillow==5.3.0
